Moved IRouteBuilder extensions into it's own class.
