+++ src/RService.IO/RServiceExtensions.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Linq;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
@@ -28,27 +26,6 @@ namespace RService.IO
 
             builder.UseRouter(routes.Build());
             return builder.UseMiddleware<RServiceMiddleware>();
-        }
-
-        public static IRouteBuilder MapRServiceIoRoute(this IRouteBuilder builder, RouteAttribute route, RequestDelegate handler)
-        {
-            var path = route.Path.Substring(1);
-
-            // ReSharper disable once SwitchStatementMissingSomeCases
-            switch (route.Verbs)
-            {
-                case RestVerbs.Any:
-                    builder.MapRoute(path, handler);
-                    break;
-                default:
-                    var verbs = route.Verbs.ToString().ToUpper()
-                        .Split(',').Select(x => x.Trim())
-                        .ToList();
-                    verbs.ForEach(verb => builder.MapVerb(verb, path, handler));
-                    break;
-            }
-
-            return builder;
         }
     }
 }
